fix: 입학탭 api v1 -> v2
