+++ apis/internal.ts
@@ -1,14 +1,14 @@
 import { FETCH_TAG_INTERNAL } from '@/constants/network';
 
-import { getRequest, putRequest } from '..';
+import { getRequest, putRequest } from '.';
 
 export const getInternal = () =>
-  getRequest('/v1/internal', undefined, {
+  getRequest('/v2/internal', undefined, {
     next: { tags: [FETCH_TAG_INTERNAL] },
   }) as Promise<{ description: string }>;
 
 export const putInternal = (description: string) =>
-  putRequest('/v1/internal', {
+  putRequest('/v2/internal', {
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({ description }),
     jsessionID: true,
