--- apis/v1/admissions/graduate/regular-admission.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getGraduateRegularAdmission = () =>
-  getRequest('/v1/admissions/graduate/regular-admission') as Promise<{ description: string }>;
