--- apis/v1/admissions/international/exchange-visiting.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getInternationalExchangeVisiting = () =>
-  getRequest('/v1/admissions/international/exchange-visiting') as Promise<{ description: string }>;
