--- apis/v1/admissions/international/graduate.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getInternationalgraduate = () =>
-  getRequest('/v1/admissions/international/graduate') as Promise<{ description: string }>;
