--- apis/v1/admissions/international/scholarships.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getInternationalScholarships = () =>
-  getRequest('/v1/admissions/international/scholarships') as Promise<{ description: string }>;
