--- apis/v1/admissions/international/undergraduate.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getInternationalUndergraduate = () =>
-  getRequest('/v1/admissions/international/undergraduate') as Promise<{ description: string }>;
