--- apis/v1/admissions/undergraduate/early-admission.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getUndergraduateEarlyAdmission = () =>
-  getRequest('/v1/admissions/undergraduate/early-admission') as Promise<{ description: string }>;
