--- apis/v1/admissions/undergraduate/regular-admission.ts
@@ -1,4 +1,0 @@
-import { getRequest } from '@/apis';
-
-export const getUndergraduateRegularAdmission = () =>
-  getRequest('/v1/admissions/undergraduate/regular-admission') as Promise<{ description: string }>;
