+++ apis/v2/admissions/graduate/regular-admission.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getGraduateRegularAdmission = () =>
+  getRequest('/v2/admissions/graduate/regular-admission') as Promise<{ description: string }>;
