+++ apis/v2/admissions/international/exchange-visiting.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getInternationalExchangeVisiting = () =>
+  getRequest('/v2/admissions/international/exchange-visiting') as Promise<{ description: string }>;
