+++ apis/v2/admissions/international/graduate.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getInternationalgraduate = () =>
+  getRequest('/v2/admissions/international/graduate') as Promise<{ description: string }>;
