+++ apis/v2/admissions/international/scholarships.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getInternationalScholarships = () =>
+  getRequest('/v2/admissions/international/scholarships') as Promise<{ description: string }>;
