+++ apis/v2/admissions/international/undergraduate.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getInternationalUndergraduate = () =>
+  getRequest('/v2/admissions/international/undergraduate') as Promise<{ description: string }>;
