+++ apis/v2/admissions/search/top.ts
@@ -2,4 +2,4 @@ import { getRequest } from '@/apis';
 import { AdmissionsSearchResult, SearchParam } from '@/apis/types/search';
 
 export const searchAdmissions = (params: SearchParam) =>
-  getRequest('/v1/admissions/search/top', params) as Promise<AdmissionsSearchResult>;
+  getRequest('/v2/admissions/search/top', params) as Promise<AdmissionsSearchResult>;
