+++ apis/v2/admissions/undergraduate/early-admission.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getUndergraduateEarlyAdmission = () =>
+  getRequest('/v2/admissions/undergraduate/early-admission') as Promise<{ description: string }>;
