+++ apis/v2/admissions/undergraduate/regular-admission.ts
@@ -1,0 +1,4 @@
+import { getRequest } from '@/apis';
+
+export const getUndergraduateRegularAdmission = () =>
+  getRequest('/v2/admissions/undergraduate/regular-admission') as Promise<{ description: string }>;
