+++ apis/v2/reservation/recurring/[id].ts
@@ -1,5 +1,5 @@
 import { deleteRequest } from '@/apis';
 
 export const deleteAllRecurringReservation = async (id: string) => {
-  await deleteRequest(`/v1/reservation/recurring/${id}`, { jsessionID: true });
+  await deleteRequest(`/v2/reservation/recurring/${id}`, { jsessionID: true });
 };
