+++ app/.internal/edit/page.tsx
@@ -1,6 +1,6 @@
 import { revalidateTag } from 'next/cache';
 
-import { getInternal, putInternal } from '@/apis/v1/internal';
+import { getInternal, putInternal } from '@/apis/internal';
 import InternalEditor from '@/app/.internal/components/InternalEditor';
 import { FETCH_TAG_INTERNAL } from '@/constants/network';
 import { redirectKo } from '@/i18n/routing';
