+++ app/.internal/page.tsx
@@ -1,4 +1,4 @@
-import { getInternal } from '@/apis/v1/internal';
+import { getInternal } from '@/apis/internal';
 import EditButton from '@/app/.internal/components/EditButton';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
