+++ app/[locale]/admissions/international/scholarships/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 
-import { getInternationalScholarships } from '@/apis/v1/admissions/international/scholarships';
+import { getInternationalScholarships } from '@/apis/v2/admissions/international/scholarships';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { internationalScholarships } from '@/constants/segmentNode';
