+++ app/[locale]/reservations/[roomType]/[roomName]/helper/ReservationCalendar.tsx
@@ -1,4 +1,4 @@
-import { getWeeklyReservation } from '@/apis/v1/reservation/week';
+import { getWeeklyReservation } from '@/apis/v2/reservation/week';
 
 import CalendarContent from './calendar/CalendarContent';
 import Toolbar from './calendar/CalendarToolbar';
