+++ app/[locale]/search/fetchContent.ts
@@ -10,7 +10,7 @@ import {
 import { SeminarPreviewList } from '@/apis/types/seminar';
 import { searchAbout } from '@/apis/v1/about/search/top';
 import { searchAcademics } from '@/apis/v1/academics/search/top';
-import { searchAdmissions } from '@/apis/v1/admissions/search/top';
+import { searchAdmissions } from '@/apis/v2/admissions/search/top';
 import { searchMember } from '@/apis/v1/member/search/top';
 import { searchNews } from '@/apis/v2/news/totalSearch';
 import { searchNotice } from '@/apis/v2/notice/totalSearch';
